move agent_message_skills type
