+++ front/types/assistant/agent_message_skills.ts
